# openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele API
  version: "1.0.0"
  description: Wedding gift management system with family tree, escrow-based pool funding, and currency snapshot locking.
servers:
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────
  # Standard responses (переиспользуемые)
  # ─────────────────────────────────────
  responses:
    BadRequest:
      description: Bad request — invalid input or business rule violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request body validation failed (Joi)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─────────────────────────────
    # Weddings
    # ─────────────────────────────
...
    PaginatedContributions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContributionResponse"
        pagination:
          $ref: "#/components/schemas/CursorPagination"

    # ─────────────────────────────
    # Family Tree
    # ─────────────────────────────
    FamilyTreeMember:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        ancestorId:
          type: integer
          nullable: true
        memberName:
          type: string
        memberEmail:
          type: string
        kinshipRank:
          type: string
          enum: [ATA_ANA, ZHIEN_ZHARAP, SHAKYRT]
        giftObligation:
          type: integer
          nullable: true
        distance:
          type: integer
        depth:
          type: integer
    FamilyTreeResponse:
      type: object
      properties:
        weddingId:
          type: integer
        tree:
          type: array
          items:
            $ref: "#/components/schemas/FamilyTreeMember"
        totalMembers:
          type: integer
    AddFamilyMemberRequest:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: integer
          example: 3
        ancestorId:
          type: integer
          nullable: true
          example: 2
        giftObligation:
          type: integer
          example: 50000
    GiftObligation:
      type: object
      properties:
        memberId:
          type: integer
        memberName:
          type: string
        memberEmail:
          type: string
        kinshipRank:
          type: string
        giftObligation:
          type: integer
          nullable: true
        contributedKzt:
          type: integer
        remainingObligation:
          type: integer
          nullable: true
        depth:
          type: integer
    GiftObligationsResponse:
      type: object
      properties:
        weddingId:
          type: integer
        obligations:
          type: array
          items:
            $ref: "#/components/schemas/GiftObligation"
        totalMembers:
          type: integer

    # ─────────────────────────────
    # Admin & Moderator
    # ─────────────────────────────
...
    PaginatedAuditLog:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        pagination:
          $ref: "#/components/schemas/CursorPagination"
...
    CursorPagination:
      type: object
      properties:
        limit:
          type: integer
          example: 10
        nextCursor:
          type: integer
          nullable: true
          example: 42
        hasNextPage:
          type: boolean
          example: true
...
paths:
...
  /family/{weddingId}/tree:
    get:
      tags: [Family Tree]
      summary: Get full family tree hierarchy with recursive structure
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: weddingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Family tree hierarchy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTreeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /family/{weddingId}/obligations:
    get:
      tags: [Family Tree]
      summary: Get gift obligations for all family members
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: weddingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gift obligations by kinship rank
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftObligationsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /family/{weddingId}/member:
    post:
      tags: [Family Tree]
      summary: Add a family member to the wedding tree
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: weddingId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFamilyMemberRequest"
      responses:
        "201":
          description: Family member added
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /family/{weddingId}/member/{memberId}:
    delete:
      tags: [Family Tree]
      summary: Remove a family member from the tree
      description: Cannot remove a member who has descendants. Re-assign their ancestorId first.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: weddingId
          required: true
          schema:
            type: integer
        - in: path
          name: memberId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed from tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /family/{weddingId}/remind:
    post:
      tags: [Family Tree]
      summary: Send gift obligation reminder emails to family members
      description: Sends reminders to all family members with outstanding gift obligations.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: weddingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reminders sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  sentCount:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────
  # Admin
  # ─────────────────────────────────────
...
  /moderator/contributions/flagged:
    get:
      tags: [Moderator]
      summary: View flagged contributions (FAILED or PENDING status)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: cursor
          schema:
            type: integer
      responses:
        "200":
          description: Paginated flagged contributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedContributions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /moderator/users/{id}/block:
    patch:
      tags: [Moderator]
      summary: Toggle block/unblock for a user
      description: Cannot block SUPER_ADMIN or yourself.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User block status toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /moderator/audit-log:
    get:
      tags: [Moderator]
      summary: View own moderator audit log
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: cursor
          schema:
            type: integer
      responses:
        "200":
          description: Paginated audit log for the current moderator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAuditLog"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────
  # Health
  # ─────────────────────────────────────
...